Inclusão de Request no Módulo de Despesas

+++ app/Http/Controllers/DespesaController.php
@@ -2,44 +2,48 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\DespesaRequest;
 use App\Models\{Despesa, Categoria};
 use Illuminate\Http\RedirectResponse;
-use Illuminate\Http\Request;
 use Illuminate\View\View;
 
 class DespesaController extends Controller
 {
     /**
      * Lista as despesas do usuário
+     *
+     * @return View
      */
     public function index(): View
     {
-        $userId = auth()->user()->id;
-        $despesas = Despesa::where('user_id', '=', $userId)->with('categoria')->get()->toArray();
+        $despesas = Despesa::where('user_id', '=', auth()->user()->id)->with('categoria')->get()->toArray();
+
         $total = array_sum(array_column($despesas, 'valor'));
+
         return view('despesas/list', ['despesas' => $despesas, 'total' => $total,  'activeDespesas' => 'active']);
     }
 
     /**
      * Renderiza o formulário de inserção de despesas
+     *
+     * @return View
      */
     public function create(): View
     {
         $categorias = Categoria::all()->toArray();
+
         return view('despesas/form', ['categorias' => $categorias, 'activeDespesas' => 'active']);
     }
 
     /**
      * Recebe os dados do formulário de inserção de receitas e salva
+     *
+     * @param DespesaRequest $request Injeção de Dependência do Laravel
+     * @return RedirectResponse
      */
-    public function store(Request $request): RedirectResponse
+    public function store(DespesaRequest $request): RedirectResponse
     {
-        $validated = $request->validate([
-            'valor' => 'required',
-            'dia_vencimento' => 'required|numeric',
-            'categoria' => 'required|string',
-            'descricao' => 'required'
-        ]);
+        $validated = $request->safe()->all();
 
         $cat = Categoria::where('nome_categoria', 'like', '%'.$validated['categoria'].'%')->first();
 
@@ -48,25 +52,27 @@ class DespesaController extends Controller
             $cat = Categoria::create(['nome_categoria' => ucfirst($validated['categoria'])]);
 
         $validated['categoria_id'] = $cat->id_categoria;
-
-        $validated['valor'] = str_replace('.', '', $validated['valor']);
-        $validated['valor'] = str_replace(',', '.', $validated['valor']);
-
+        $validated['valor'] = str_replace(',', '.', str_replace('.', '', $validated['valor']));
         $validated['user_id'] = auth()->user()->id;
 
-        $despesa = new Despesa($validated);
-
-        $despesa->save();
+        Despesa::create($validated);
 
         return redirect(route('despesas.index'))->with('msg', ['type' => 'success', 'msg' => 'Despesa cadastrada!']);
     }
 
     /**
      * Renderiza o formulário para editar uma despesa
+     *
+     * @param int $id_despesa Código da Despesa
+     * @return RedirectResponse|View
      */
-    public function edit($id): View
+    public function edit(int $id_despesa): RedirectResponse|View
     {
-        $despesa = Despesa::where('id_despesa', '=', $id)->with('categoria')->first()->toArray();
+        $despesa = Despesa::where('id_despesa', '=', $id_despesa)->with('categoria')->first();
+
+        if(empty($despesa))
+            return redirect(route('despesas.index'))->with('msg', ['type' => 'danger', 'msg' => 'Despesa informada não encontrada']);
+
         $despesa['valor'] = number_format($despesa['valor'], 2, ',', '.');
 
         $categorias = Categoria::all();
@@ -76,26 +82,30 @@ class DespesaController extends Controller
 
     /**
      * Recebe os dados do formulário de edição de despesa e salva
+     *
+     * @param int $id_despesa Código da Despesa
+     * @param DespesaRequest $request Injeção de Dependência do Laravel
+     * @return RedirectResponse
      */
-    public function update($id, Request $request): RedirectResponse
+    public function update(int $id_despesa, DespesaRequest $request): RedirectResponse
     {
-        $validated = $request->validate([
-            'valor' => 'required',
-            'dia_vencimento' => 'required|numeric',
-            'categoria_id' => 'required|numeric',
-            'descricao' => 'required'
-        ]);
+        $despesa = Despesa::where('id_despesa', '=', $id_despesa)->where('user_id', '=', auth()->user()->id)->with('categoria')->first();
 
-        $validated['valor'] = str_replace('.', '', $validated['valor']);
-        $validated['valor'] = str_replace(',', '.', $validated['valor']);
-        $validated['user_id'] = auth()->user()->id;
+        if(empty($despesa))
+            return redirect(route('despesas.index'))->with('msg', ['type' => 'danger', 'msg' => 'Despesa informada não encontrada']);
 
-        $despesa = Despesa::findOrFail($id);
-        $despesa->valor = $validated['valor'];
+        $validated = $request->safe()->all();
+
+        //Caso a categoria editada não exista, insere na tabela de categorias
+        $cat = Categoria::where('nome_categoria', 'like', '%'. $validated['categoria'] .'%')->first();
+
+        if(empty($cat))
+            $cat = Categoria::create(['nome_categoria' => ucfirst($validated['categoria'])]);
+
+        $despesa->valor = str_replace(',', '.', str_replace('.', '', $validated['valor']));
         $despesa->dia_vencimento = $validated['dia_vencimento'];
-        $despesa->categoria_id = $validated['categoria_id'];
         $despesa->descricao = $validated['descricao'];
-        $despesa->user_id = $validated['user_id'];
+        $despesa->categoria_id = $cat->id_categoria;
 
         $despesa->update();
 
@@ -104,10 +114,16 @@ class DespesaController extends Controller
 
     /**
      * Remove a despesa informada
+     *
+     * @param int $id_despesa Código da Despesa
+     * @return RedirectResponse
      */
-    public function destroy($id): RedirectResponse
+    public function destroy(int $id_despesa): RedirectResponse
     {
-        $despesa = Despesa::findOrFail($id);
+        $despesa = Despesa::where('id_despesa', '=', $id_despesa)->where('user_id', '=', auth()->user()->id)->with('categoria')->first();
+
+        if(empty($despesa))
+            return redirect(route('despesas.index'))->with('msg', ['type' => 'danger', 'msg' => 'Despesa informada não encontrada']);
 
         $despesa->delete();
